Duplicate interfaces refactored
