+++ tests/pages/api/apiUtils/userData.test.ts
@@ -26,12 +26,10 @@ import {
     MULTIPLE_OPERATOR_ATTRIBUTE,
     SCHOOL_FARE_TYPE_ATTRIBUTE,
 } from '../../../../src/constants/index';
-
 import {
     defaultSalesOfferPackageOne,
     defaultSalesOfferPackageTwo,
 } from '../../../../src/pages/selectSalesOfferPackage';
-
 import {
     getMockRequestAndResponse,
     expectedSingleTicket,
@@ -55,8 +53,6 @@ import {
 } from '../../../testData/mockData';
 import * as s3 from '../../../../src/data/s3';
 import * as auroradb from '../../../../src/data/auroradb';
-
-import { FareZone } from '../../../../src/pages/api/csvZoneUpload';
 import { Operator, MultipleProductAttribute } from '../../../../src/interfaces';
 
 describe('userData', () => {
@@ -255,13 +251,12 @@ describe('userData', () => {
             ['period', expectedPeriodGeoZoneTicketWithMultipleProducts],
             ['multiOperator', expectedMultiOperatorGeoZoneTicketWithMultipleProducts],
         ])('should return a %s geoZoneTicket object', async (fareType, expectedJson) => {
-            const mockFareZoneAttribute: FareZone = { fareZoneName: 'Green Lane Shops' };
             const { req, res } = getMockRequestAndResponse({
                 cookieValues: {},
                 session: {
                     [TIME_RESTRICTIONS_DEFINITION_ATTRIBUTE]: mockTimeRestriction,
                     [FARE_TYPE_ATTRIBUTE]: { fareType },
-                    [FARE_ZONE_ATTRIBUTE]: mockFareZoneAttribute,
+                    [FARE_ZONE_ATTRIBUTE]: { fareZoneName: 'Green Lane Shops' },
                     [MULTIPLE_PRODUCT_ATTRIBUTE]: {
                         products: [
                             {
@@ -516,7 +511,6 @@ describe('userData', () => {
         });
 
         it('should return a SchemeOperatorTicket object', async () => {
-            const mockFareZoneAttribute: FareZone = { fareZoneName: 'Green Lane Shops' };
             const { req, res } = getMockRequestAndResponse({
                 cookieValues: {
                     operator: { name: 'SCHEME_OPERATOR', region: 'SCHEME_REGION', nocCode: 'TESTSCHEME' },
@@ -525,7 +519,7 @@ describe('userData', () => {
                 session: {
                     [TIME_RESTRICTIONS_DEFINITION_ATTRIBUTE]: mockTimeRestriction,
                     [FARE_TYPE_ATTRIBUTE]: { fareType: 'multiOperator' },
-                    [FARE_ZONE_ATTRIBUTE]: mockFareZoneAttribute,
+                    [FARE_ZONE_ATTRIBUTE]: { fareZoneName: 'Green Lane Shops' },
                     [MULTIPLE_PRODUCT_ATTRIBUTE]: {
                         products: [
                             {
